requirements.txt
streamlit>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
pathlib>=1.0.1 